refactor: 💡 Move global setup to root
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-14777
