+++ e2e-tests/admin/tests/credential-store-vault-ent.spec.js
@@ -6,7 +6,7 @@
 import { test } from '../playwright.config.js';
 import { execSync } from 'child_process';
 
-import { authenticatedState } from '../global-setup.js';
+import { authenticatedState } from '../../global-setup.js';
 import {
   authenticateBoundaryCli,
   authorizeSessionByTargetIdCli,
