+++ e2e-tests/admin/tests/target-ent.spec.js
@@ -6,7 +6,7 @@
 import { test } from '../playwright.config.js';
 import { expect } from '@playwright/test';
 
-import { authenticatedState } from '../global-setup.js';
+import { authenticatedState } from '../../global-setup.js';
 import {
   authenticateBoundaryCli,
   checkBoundaryCli,
