+++ e2e-tests/admin/tests/worker-tags.spec.js
@@ -7,7 +7,7 @@ import { test } from '../playwright.config.js';
 import { expect } from '@playwright/test';
 import { customAlphabet } from 'nanoid';
 
-import { authenticatedState } from '../global-setup.js';
+import { authenticatedState } from '../../global-setup.js';
 import { WorkersPage } from '../pages/workers.js';
 
 test.use({ storageState: authenticatedState });
