+++ e2e-tests/admin/tests/worker.spec.js
@@ -6,7 +6,7 @@
 import { test } from '../playwright.config.js';
 import { expect } from '@playwright/test';
 
-import { authenticatedState } from '../global-setup.js';
+import { authenticatedState } from '../../global-setup.js';
 
 test.use({ storageState: authenticatedState });
